refactor: add messages module
