+++ src/modules/dir.js
@@ -17,12 +17,13 @@ const setHomeDir = () => {
   process.chdir(homedir())
 }
 
-const printHomeDir = () => {
-  console.log(`You are currently in ${homedir()}`)
-}
+// const printHomeDir = () => {
+//   console.log(`You are currently in ${homedir()}`)
+// }
 
-const printCurrentDir = () => {
-  console.log(`You are currently in ${process.cwd()}`)
-}
+// const printCurrentDir = () => {
+//   console.log(`You are currently in ${process.cwd()}`)
+// }
 
-export { printHomeDir, printCurrentDir, setHomeDir, changeDir, goToUpperDir }
+export { setHomeDir, changeDir, goToUpperDir }
+// export { printHomeDir, printCurrentDir, setHomeDir, changeDir, goToUpperDir }
